Added default user agent header.
From the GitHub API documentation: 
All API requests MUST include a valid User Agent string. Requests with no User Agent string will be rejected. If you are hitting the API without authentication, we ask that you add some kind of identification to the UA header value. This is so we can contact you if there are problems.

The default UA does no harm for other providers.